✨ Added RenderViewManager
