+++ appcues/src/main/java/com/appcues/trait/appcues/TooltipTrait.kt
@@ -52,7 +52,7 @@ import com.appcues.trait.extensions.getTooltipPointerPosition
 import com.appcues.trait.extensions.rememberDpStepAnimation
 import com.appcues.trait.extensions.rememberFloatStepAnimation
 import com.appcues.trait.extensions.rememberTargetRectangleInfo
-import com.appcues.ui.AppcuesOverlayViewManager
+import com.appcues.ui.RenderViewManager
 import com.appcues.ui.composables.LocalAppcuesStepMetadata
 import com.appcues.ui.composables.rememberAppcuesContentVisibility
 import com.appcues.ui.extensions.coloredShadowPath
@@ -101,7 +101,7 @@ internal class TooltipTrait(
     }
 
     override fun present() {
-        val success = AppcuesOverlayViewManager(scope, renderContext).start()
+        val success = RenderViewManager(scope, renderContext).start()
 
         if (!success) {
             throw AppcuesTraitException("unable to create tooltip overlay view")
